Add files via upload
Not recommended - Optional file to hide a lot of the selections so they only appear in 'Random'. This does make the menu more manageable and still displays 3-stage evolutions and trios. 
File must be renamed to pokecards.lua and then put into the mod replacing the existing pokecards.lua file.